feat: добавить модуль ewm

+++ ewm/src/main/java/practicum/model/UpdateCompilationRequest.java
@@ -3,7 +3,7 @@ package practicum.model;
 import java.util.List;
 
 public class UpdateCompilationRequest {
-    List<Integer> events;
+    List<Long> events;
     Boolean pinned;
     String title;
 }
